Move memory components into dinocpu.memory package
