+++ src/main/scala/components/memory/memory-combin-ports.scala
@@ -1,10 +1,10 @@
 // Combinational/'synchronous' memory module
 
-package dinocpu
+package dinocpu.memory
 
 import chisel3._
 import chisel3.util._
-import dinocpu.MemoryOperation._
+import dinocpu.memory.MemoryOperation._
 
 /**
   * The instruction memory port.
