+++ src/main/scala/components/memory/memory-operations.scala
@@ -1,6 +1,6 @@
 // Asynchronous memory module
 
-package dinocpu
+package dinocpu.memory
 
 import chisel3._
 import chisel3.experimental.ChiselEnum
