+++ src/main/scala/configuration.scala
@@ -4,6 +4,7 @@ package dinocpu
 
 // For file length
 import java.io.{File}
+import dinocpu.memory._
 
 /**
  * This class configures all of the dinocpus. It takes parameters for the type of CPU model
