+++ src/test/scala/components/CombinMemoryUnitTest.scala
@@ -4,8 +4,7 @@ package dinocpu
 
 import chisel3._
 import chisel3.iotesters.{ChiselFlatSpec, Driver, PeekPokeTester}
-
-
+import dinocpu.memory._
 
 /**
   * Testing harness for the combinational memory
