+++ src/test/scala/components/NonCombinMemoryUnitTest.scala
@@ -4,6 +4,7 @@ package dinocpu
 
 import chisel3._
 import chisel3.iotesters.{ChiselFlatSpec, Driver, PeekPokeTester}
+import dinocpu.memory._
 
 // To generate random latencies
 import scala.util.Random
